fix estimation

+++ monitor/src/main.cpp
@@ -324,7 +324,7 @@ void renderHeader() {
   // https://www.powersol.com.ua/ibp/kak-rasschitat-vremya-avtonomnoj-raboty-ibp/#:~:text=%D0%92%D0%BE%D0%BE%D1%80%D1%83%D0%B6%D0%B8%D0%B2%D1%88%D0%B8%D1%81%D1%8C%20%D0%BA%D0%B0%D0%BB%D1%8C%D0%BA%D1%83%D0%BB%D1%8F%D1%82%D0%BE%D1%80%D0%BE%D0%BC%2C%20%D1%80%D0%B0%D1%81%D1%87%D0%B5%D1%82%20%D0%BC%D0%BE%D0%B6%D0%BD%D0%BE%20%D0%BE%D1%81%D1%83%D1%89%D0%B5%D1%81%D1%82%D0%B2%D0%B8%D1%82%D1%8C,%D0%BF%D0%BE%D0%BB%D1%83%D1%87%D0%B0%D0%B5%D0%BC%20%3D%20%D0%9A%D0%BE%D0%BB%D0%B8%D1%87%D0%B5%D1%81%D1%82%D0%B2%D0%BE%20%D1%87%D0%B0%D1%81%D0%BE%D0%B2%20%D0%BD%D0%B5%D0%BF%D1%80%D0%B5%D1%80%D1%8B%D0%B2%D0%BD%D0%BE%D0%B9%20%D1%80%D0%B0%D0%B1%D0%BE%D1%82%D1%8B.
   // estimate working time
   if (vin < 12 && bat_count > 0) {
-    int et_total = 60.0 * bus * batcap_total * K * (bat_level(bus) / 100.0) / pwr;
+    long et_total = 60.0 * bus * batcap_total * K * (bat_level(bus) / 100.0) / pwr;
     int et_d = et_total / (60 * 24);
     int et_h = (et_total % (60 * 24)) / 60;
     int et_m = et_total % (60 * 24) % 60;
